requirements.txt
# Azure Functions
azure-functions==1.18.0

# Slack Bot Framework
slack-bolt==1.18.1
slack-sdk==3.26.2

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Date/time utilities
python-dateutil==2.8.2

# Azure Service Bus
azure-servicebus==7.12.2

# YAML support
PyYAML==6.0.1

# System information
psutil==5.9.8

# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry
